+++ R/data-dataset.R
@@ -9,6 +9,13 @@
 #' @param dataset Selected dataset
 #' @param package Name of an R package containing datasets
 #'
+#' @block dataset block
+#' @blockDescr Choose a dataset from a package
+#' @blockCategory input
+#' @blockIcon database
+#' @blockParam dataset Selects the dataset to use.
+#' @blockParamExample dataset iris
+#'
 #' @rdname new_data_block
 #' @export
 new_dataset_block <- function(dataset = character(), package = "datasets",